requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
blas=1.0=openblas
bottleneck=1.3.5=py38h67323c0_0
brotli=1.0.9=hca72f7f_7
brotli-bin=1.0.9=hca72f7f_7
ca-certificates=2023.01.10=hecd8cb5_0
cairo=1.16.0=h3ce6f7e_3
certifi=2022.12.7=py38hecd8cb5_0
cffi=1.15.1=py38h6c40b1e_3
contourpy=1.0.5=py38haf03e11_0
cycler=0.11.0=pyhd3eb1b0_0
eigen=3.3.7=haf03e11_1
expat=2.4.9=he9d5cce_0
flit-core=3.6.0=pyhd3eb1b0_0
fontconfig=2.14.1=hedf32ac_1
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=hd8bbffd_0
future=0.18.2=py38_1
gettext=0.21.0=h7535e17_0
giflib=5.2.1=h6c40b1e_1
glib=2.69.1=hfff2838_2
graphite2=1.3.14=he9d5cce_1
gst-plugins-base=1.14.0=h4180768_2
gstreamer=1.14.0=h0fc69c2_2
harfbuzz=4.3.0=h3f129a2_0
hdf5=1.10.6=h10fe05b_1
icu=58.2=h0a44026_3
intel-openmp=2021.4.0=hecd8cb5_3538
jpeg=9e=hca72f7f_0
kiwisolver=1.4.4=py38hcec6c5f_0
krb5=1.19.4=hdba6334_0
lcms2=2.12=hf1fd2bf_0
lerc=3.0=he9d5cce_0
libbrotlicommon=1.0.9=hca72f7f_7
libbrotlidec=1.0.9=hca72f7f_7
libbrotlienc=1.0.9=hca72f7f_7
libclang=12.0.0=default_hbc2896b_2
libcxx=14.0.6=h9765a3e_0
libdeflate=1.8=h9ed2024_5
libedit=3.1.20221030=h6c40b1e_0
libffi=3.4.2=hecd8cb5_6
libgfortran=5.0.0=11_3_0_hecd8cb5_28
libgfortran5=11.3.0=h9dfd629_28
libiconv=1.16=hca72f7f_2
libllvm12=12.0.0=h9b2ccf5_3
libopenblas=0.3.21=h54e7dc3_0
libpng=1.6.37=ha441bb4_0
libpq=12.9=h1c9f633_3
libprotobuf=3.20.3=hfff2838_0
libtiff=4.5.0=h2cd0358_0
libuv=1.40.0=haf1e3a3_0
libwebp=1.2.4=h56c3ce4_0
libwebp-base=1.2.4=hca72f7f_0
libxml2=2.9.14=hbf8cd5e_0
libxslt=1.1.35=h5b33f42_0
llvm-openmp=14.0.6=h0dcd299_0
lz4-c=1.9.4=hcec6c5f_0
matplotlib=3.6.2=py38hecd8cb5_0
matplotlib-base=3.6.2=py38h220de94_0
mkl=2021.4.0=hecd8cb5_637
mkl-service=2.4.0=py38h9ed2024_0
munkres=1.1.4=py_0
ncurses=6.4=hcec6c5f_0
ninja=1.10.2=hecd8cb5_5
ninja-base=1.10.2=haf03e11_5
nspr=4.33=he9d5cce_0
nss=3.74=h47edf6a_0
numexpr=2.8.4=py38h57a7bef_0
numpy=1.23.5=py38h57a7bef_0
numpy-base=1.23.5=py38hc93c6d9_0
opencv=4.6.0=py38h0e6eb04_2
openjpeg=2.4.0=h66ea3da_0
openssl=1.1.1s=hca72f7f_0
packaging=22.0=py38hecd8cb5_0
pandas=1.5.2=py38h07fba90_0
pcre=8.45=h23ab428_0
pillow=9.3.0=py38h81888ad_1
pip=22.3.1=py38hecd8cb5_0
pixman=0.40.0=h9ed2024_1
pycparser=2.21=pyhd3eb1b0_0
pyparsing=3.0.9=py38hecd8cb5_0
python=3.8.16=h218abb5_2
python-dateutil=2.8.2=pyhd3eb1b0_0
pytorch=1.12.1=cpu_py38h64f2f56_1
pytz=2022.7=py38hecd8cb5_0
pyyaml=6.0=py38h6c40b1e_1
qt-main=5.15.2=h719ae48_7
qt-webengine=5.15.9=h90a370e_4
qtwebkit=5.212=h24dc246_4
readline=8.2=hca72f7f_0
setuptools=65.6.3=py38hecd8cb5_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.40.1=h880c91c_0
tk=8.6.12=h5d9f67b_0
tornado=6.2=py38hca72f7f_0
typing-extensions=4.4.0=py38hecd8cb5_0
typing_extensions=4.4.0=py38hecd8cb5_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.10=h6c40b1e_1
yaml=0.2.5=haf1e3a3_0
zlib=1.2.13=h4dc903c_0
zstd=1.5.2=hcb37349_0
